reflectFlowのユニットフローを改善
